Move care-pathway-service onto loco-rs 1.0.1
Third crate on the family-wide loco-rs 1.0.1 migration. Same base recipe
as authentication-service and organization-service, plus two things
specific to this crate:

- sea-orm 2.0 dropped `DatabaseConnection::Disconnected`, which
  src/compliance/disclosure.rs's test used as a "would fail if touched"
  stand-in to prove the read-audit no-op path never reaches the database
  when auditing is off. Replaced with a `MockDatabase` carrying no
  queued results — it errors identically on the first real query — added
  as a `mock`-feature dev-dependency so it never reaches a release build.
- The 64-bit PkAuto width change reaches further here than in
  organization: beyond the three entities themselves, care_pathways'
  audit hash-chain carries row ids as first-class values in its own
  types (ChainBreak.id, Checkpoint.anchor_id), which had to widen too.
  The five instances/outcomes tables and event_outbox stay i32/SERIAL —
  same reason as organization's event_outbox: their migrations write raw
  SQL rather than going through the loco schema DSL.

No SOUP register change needed (compliance/soup.tsv) — loco-rs and
sea-orm were already annotated from S-2; this is a version bump of
existing dependencies, not a new one.

Verified: cargo build --all-targets, fmt, clippy -D warnings, cargo test
(239 passed, including the SOUP gate) clean; the full DB-gated suite (46
tests, unchanged count) green against a freshly migrated Postgres 18.

+++ care-pathway/care-pathway-service-with-loco/src/compliance/checkpoint.rs
@@ -61,7 +61,7 @@ pub struct Checkpoint {
     /// Format version.
     pub version: String,
     /// The anchor row's `id` — the position this checkpoint describes.
-    pub anchor_id: i32,
+    pub anchor_id: i64,
     /// The anchor row's stored SHA-256 chain hash.
     pub head: String,
     /// How many rows stood at or before `anchor_id` when this was taken.
@@ -91,7 +91,7 @@ pub enum CheckpointVerdict {
     /// The anchor row is gone: rows have been deleted.
     AnchorMissing {
         /// The `id` that should have been present.
-        anchor_id: i32,
+        anchor_id: i64,
     },
     /// The anchor survived but no longer hashes to the recorded head.
     HeadChanged {
