+++ care-pathway/care-pathway-service-with-loco/src/compliance/record_integrity.rs
@@ -316,7 +316,7 @@ mod tests {
     }
 
     /// A stored row with a correct hash.
-    fn row(id: i32, name: &str) -> care_pathways::Model {
+    fn row(id: i64, name: &str) -> care_pathways::Model {
         let mut model = care_pathways::Model {
             created_at: at(1_700_000_000_000_000),
             updated_at: at(1_700_000_000_000_000),
